# docs/swagger.yaml
openapi: 3.0.0
info:
  title: ZRS Cars Trading CRM API
  description: |
    Comprehensive CRM API for ZRS Cars Trading managing purchases, inventory, sales, investors, and customer service.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Roles
    - **Admin**: Full system access
    - **Manager**: Lead management, PO creation, inventory operations
    - **Investor**: Read-only access to portfolio
    
  version: 1.0.0
  contact:
    email: dev@zrscarstrading.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.zrscarstrading.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /webhooks/docusign:
    post:
      tags: [Webhooks]
      summary: DocuSign webhook callback
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed